backups/20250625_113403/message-consumer/requirements.txt
flask==2.0.1
pika==1.3.0
kafka-python==2.0.2
prometheus-client==0.16.0
prometheus-flask-exporter==0.21.0
requests==2.31.0
flask==2.0.1
werkzeug==2.0.3